Fix build to work properly
The build directories are now nicely linked up to the correct output path
Now 2 folders will be generated, 32bit and 64bit, with a release build inside of them as well as a subdirectory for the debug build. The debug build will not work as is as it can not find the Data folder. Make sure to drag the debug folder out before running it.